requirements.txt
aiohttp>=3.8.0
beautifulsoup4>=4.9.3
fastapi>=0.68.0
forex-python>=1.8.0
numpy>=1.21.0
openai>=1.0.0
Pillow>=9.0.0
pydantic-settings>=2.0.0
pycoingecko>=3.0.0
redis>=4.0.0
spacy>=3.0.0
structlog>=22.1.0
tavily-python>=0.1.9
textblob>=0.15.3
uvicorn>=0.15.0
yfinance>=0.2.0
nest-asyncio>=1.5.0
replicate>=0.20.0
